Initial run of immaculate
- Hasn't fully been run, certain files cause exceptions

+++ src/main/java/com/simibubi/create/content/logistics/chute/ChuteItemHandler.java
@@ -3,54 +3,49 @@ package com.simibubi.create.content.logistics.chute;
 import com.simibubi.create.foundation.item.ItemHelper;
 
 import net.minecraft.world.item.ItemStack;
-
 import net.neoforged.neoforge.items.IItemHandler;
 
 public class ChuteItemHandler implements IItemHandler {
 
-	private ChuteBlockEntity blockEntity;
+    private final ChuteBlockEntity blockEntity;
 
-	public ChuteItemHandler(ChuteBlockEntity be) {
-		this.blockEntity = be;
-	}
+    public ChuteItemHandler(ChuteBlockEntity be) {
+        this.blockEntity = be;
+    }
 
-	@Override
-	public int getSlots() {
-		return 1;
-	}
+    @Override
+    public int getSlots() {
+        return 1;
+    }
 
-	@Override
-	public ItemStack getStackInSlot(int slot) {
-		return blockEntity.item;
-	}
+    @Override
+    public ItemStack getStackInSlot(int slot) {
+        return blockEntity.item;
+    }
 
-	@Override
-	public ItemStack insertItem(int slot, ItemStack stack, boolean simulate) {
-		if (!blockEntity.canAcceptItem(stack))
-			return stack;
-		ItemStack remainder = ItemHelper.limitCountToMaxStackSize(stack, simulate);
-		if (!simulate)
-			blockEntity.setItem(stack);
-		return remainder;
-	}
+    @Override
+    public ItemStack insertItem(int slot, ItemStack stack, boolean simulate) {
+        if (!blockEntity.canAcceptItem(stack)) return stack;
+        ItemStack remainder = ItemHelper.limitCountToMaxStackSize(stack, simulate);
+        if (!simulate) blockEntity.setItem(stack);
+        return remainder;
+    }
 
-	@Override
-	public ItemStack extractItem(int slot, int amount, boolean simulate) {
-		ItemStack remainder = blockEntity.item.copy();
-		ItemStack split = remainder.split(amount);
-		if (!simulate)
-			blockEntity.setItem(remainder);
-		return split;
-	}
+    @Override
+    public ItemStack extractItem(int slot, int amount, boolean simulate) {
+        ItemStack remainder = blockEntity.item.copy();
+        ItemStack split = remainder.split(amount);
+        if (!simulate) blockEntity.setItem(remainder);
+        return split;
+    }
 
-	@Override
-	public int getSlotLimit(int slot) {
-		return Math.min(getStackInSlot(slot).getMaxStackSize(), 64);
-	}
+    @Override
+    public int getSlotLimit(int slot) {
+        return Math.min(getStackInSlot(slot).getMaxStackSize(), 64);
+    }
 
-	@Override
-	public boolean isItemValid(int slot, ItemStack stack) {
-		return true;
-	}
-
+    @Override
+    public boolean isItemValid(int slot, ItemStack stack) {
+        return true;
+    }
 }
